requirements.txt
sanic==19.12.2
newspaper3k==0.2.8